+++ src/Mancare.cpp
@@ -1,4 +1,4 @@
 #include "Mancare.hpp"
 
-Mancare::Mancare(const std::string& nume, float greutate, int raritate, int energie) : Loot(nume, greutate, raritate), energieRecuperata(energie) {}
+Mancare::Mancare(const std::string& nume, double greutate, int raritate, int energie) : Loot(nume, greutate, raritate), energieRecuperata(energie) {}
 Mancare::~Mancare() = default;
