+++ src/Material.cpp
@@ -1,6 +1,6 @@
 #include "Material.hpp"
 
 
-Material::Material(const std::string& nume, float greutate, int raritate) : Loot(nume, greutate, raritate) {}
+Material::Material(const std::string& nume, double greutate, int raritate) : Loot(nume, greutate, raritate) {}
 Material::~Material() = default;
